refac: moved `array_to_json()` functions to `utils/serializers.py`
