+++ src/careamics/utils/__init__.py
@@ -8,6 +8,7 @@ __all__ = [
     "get_logger",
     "get_careamics_home",
     "autocorrelation",
+    "array_to_json",
 ]
 
 
@@ -17,3 +18,4 @@ from .context import cwd, get_careamics_home
 from .logging import get_logger
 from .path_utils import check_path_exists
 from .ram import get_ram_size
+from .serializers import array_to_json
